requirements.txt
fastapi==0.93.0
uvicorn==0.20.0
psycopg2-binary==2.9.5
matplotlib==3.7.1
